git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@3641 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/gui/sql/SplitPaneExpander.java
@@ -23,6 +23,7 @@
 package workbench.gui.sql;
 
 import javax.swing.JSplitPane;
+
 import workbench.gui.WbSwingUtilities;
 
 /**
@@ -49,7 +50,7 @@ public class SplitPaneExpander
 		}
 		else
 		{
-			int newLoc = (int)(this.contentPanel.getHeight() / 2);
+			int newLoc = this.contentPanel.getHeight() / 2;
 			this.contentPanel.setDividerLocation(newLoc);
 		}
 		this.lastDivider = -1;
